nova-chatbot/requirements.txt
streamlit>=1.39.0
groq==0.4.2
httpx==0.27.0
python-dotenv>=1.0.0
